requirements.txt
# asgiref==3.9.1
# Django==3.2.20
# djangorestframework>=3.14.0
# djangorestframework-simplejwt>=5.2.0
# django-cors-headers>=4.0.0
# python-decouple>=3.8
# pymongo==3.12.3
# djongo==1.3.6
# dnspython>=2.3.0
# sendgrid>=6.10.0
# django-sendgrid-v5>=1.2.0
# django-extensions>=3.2.0
# motor>=3.1.1
# pymongo[srv]>=4.3.3
# python-dotenv>=1.0.0

# Core Django packages
Django==4.2.7
djongo==1.3.6

# Database
pymongo==3.12.3
dnspython==2.4.2

# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# CORS handling
django-cors-headers==4.3.1

# Email integration
sendgrid==6.10.0

# Password validation and security
django-extensions==3.2.3

# Environment variables
python-decouple==3.8

# Development tools (optional)
django-debug-toolbar==4.2.0

# Production server (optional)
gunicorn==21.2.0
whitenoise==6.6.0

